[WebAudio DevTools] Add new graph events to browser_protocol.pdl
This is a follow-up CL of:
https://chromium-review.googlesource.com/c/chromium/src/+/1726957

This CL introduces the actual event hooks between the WebAudio module
and the devtool frontend.

This CL completes the backend for the graph visualizer. On top of
this change, the graph visualizer is ready for the implementation.

Bug: 970898
Change-Id: I76fae9d19d501158fac3c13cb6226d72fccb798c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728265
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Erik Luo <luoe@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685048}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87be32736e44c5fa9ca6b34a31cdff17e0e2abec