Use Karma for unit tests
We currently use layout tests for all of the DevTools front end, which requires
that all output be serializable. This CL introduces unit tests run via Karma,
Mocha, and Chai, as well as a minimally-changed port of the utilities layout
tests in the Chromium src.

The expectation is that these tests will be run locally, i.e. npx karma start.

Change-Id: If18e8a0bdd810903e82f76f478bf16dbbb28859e