Update Closure compiler to v20190729
The new compiler caught a lot of pre-existing issues in the codebase.
Sadly, the old compiler version was not smart enough to understand the
new changes. Therefore, the changes have be included in the same CL as
the compiler update.

Most of the changes are related to better handling of prototype and
class inheritance, as well as handling of null/undefined tracking.

Fix `git cl format`

If you would run `git cl format` in this repository, it would fail,
because neither buildtools nor clang-format were added as a DEPS. This
CL adds the necessary configuration (copied from chromium/src/DEPS) to
make the command work.

Bug: 991710
Change-Id: I53e7b95239b262d66c8cc1f70f274054ef29d0d9