Update patch set 7
Patch Set 7:

(1 comment)

So overall I have a little reticence here. We're special-casing and baking in the trace categories for a variant, and it's conceivable we could have several of those. It seems like a better longer term solution would be to support toggling categories (similar to the way trace viewer does it) with perhaps some presets. That said, since we don't have any immediate plans to offer that I can see this is a pragmatic solution.

Patch-set: 7
CC: Gerrit User 1189020 <1189020@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>