Update patch set 1
Patch Set 1:

(2 comments)

Could you include the link to the old layout test?

Also, I was under the impression we would also tidy up the actual test. The test that is being ported is not readable and is unclear what it is testing. It would be better to port this to clean Karma tests that are descriptive, succinct and follow testing best practices.

Lastly, it would be great to also have some new tests for behavior that we deem untested. Granted, it would be nice to have a coverage report here, but hopefully we can generate that soon.

Regarding the commit message, please take a look at https://chris.beams.io/posts/git-commit/ which is a great guide on how to write commit messages for yourself when you are reading your own logs.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>