Update patch set 1
Patch Set 1:

(4 comments)

This is a great step forward! The tests are already a lot more readable and cleaner in terms of setup. However, most of the tests still have a lot of asserts. It would be great to split that up and only assert 1 thing at time. Note that this *could* mean there is only 1 assert.equal per testcase, but this is not necessary. It can mean you `assert.equal` 2-3 things per test, as you need to verify both the nodeName and className.

If I understand the tests correctly, they are testing multiple scenarios. The goal would be to have 1 testcase per scenario with a descriptive title as well as relevant asserts.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>