Update patch set 4
Patch Set 4: Code-Review+1

LGTM, however it needs to be a CL on Chromium rather than devtools/devtools-frontend (I didn't see this CL was targeting a different repo up to this point). Do you have a Chromium checkout you can make the CL on?

As a follow-up change to this, it would be great if we can instruct Chromium DevTools presubmit to also run these Karma tests. Let me know if we need to do that work or if you are able to do so.

Patch-set: 4
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1