DevTools: Mark audits progress bar with role="progressbar"
This change give the audits status progress bar role="progressbar"
with values marking it for assistive technology like screen readers.

To support this change UI.ARIAUtils.setProgressBarValue was added which
can optionally set aria-valuetext. The audits status has a value that
matches the index of the phase we are in.

This meets the following WCAG success criteria
1.3.1 https://www.w3.org/WAI/WCAG21/quickref/#info-and-relationships
4.1.2 https://www.w3.org/WAI/WCAG21/quickref/#name-role-value

Note on testing: The progress bar value is already read out by the
aria-live="status" on most screen readers like NVDA on windows.

Change-Id: I5fead17a7bd859129f1148b1107a319ea29af8bb