DevTools: Defensive guard of Elements search results
There is a race condition between dismissing search and stepping
search which occasionally results in an unhandled exception having
been raised. Most of the code paths dealing with search in Elements
panel already defensively code against this; this change adds similar
tests in the unguarded code paths to what already exists, and alters
the declaration of the search-results variable so that Closure would
have correctly detected this condition previously.

Change-Id: I929b0851a5770fb26b1f73c682b86a7686a7ee80