DevTools: Don't scrollIntoView non-Element nodes
During search in the Elements pane, it's possible and likely
common that users might enter some search term that would match nodes
which may not be scrolled-to, such as CDATA nodes, comments, or other
types. That's because, as opposed to the "Scroll Into View" context
menu item which guards on the types of nodes, the searcher just
indiscriminately tells the page to scroll the node in. Rather than
addressing that issue in the caller, this change guards that from
happening by calling to the page.

Also: Presubmit check wants newline at EOF.
Change-Id: I6c6501b7a88c7de894a97957029be268aaa43753