Devtools: Add aria-label for tab panel view container
Issue: tab panel relies on its view container to set the default focus on initial
load or capture focus in the case of losing focus, since it doesn't have an aria
label, screen reader reads out the whole content of the tab panel.

This change add an aria-label for each view container as "${tab-title} panel".


Bug: 963183
Change-Id: Id17b9c259b35c61a1f6cf0035ef206b2c2f2e2d1