Devtools: Make slider in Color Picker keyboard accessible
This change makes Hue Slider and Alpha Slider in Color Picker both keyboard accessible.

Keyboard Accessibility:
- Arrow Right, Arrow Up moves the slider towards right and increases the value
- Arrow Left, Arrow Down moves the slider towards left and decreases the value

In positionHue and positionAlpha
hueAlphaLeft = this._hueElement.getBoundingClientRect().left instead of this._hueElement.totalOffsetLeft() because it was returning 0 in keydown event and worked only for mouse event.
The value returned by this._hueElement.getBoundingClientRect().left and this._hueElement.totalOffsetLeft() is same in mousedown event. So, this._hueElement.getBoundingClientRect().left
is used as it works for both mousedown and keydown events.

Gif showing the changes https://imgur.com/a/jObra8z

This was reviewed earlier https://chromium-review.googlesource.com/c/chromium/src/+/1703062 which has been split into
1. Devtools: Make slider in Color Picker keyboard accessible
2. Devtools: Make slider in Color Picker screen-reader accessible

In original change keyboard accessibility had dependency on installDragHandle in this change it doesn't have that dependency.

Bug: 963183
Change-Id: Ice1a2a197ad1b84ce68026bf1d5e7016ff497c79