Changes Drawer: Fix bug to show only changed lines in Changes Drawer when the content is Pretty Printed
Earlier, when user made change and pressed 'Pretty Print' the changes made by 'Pretty Print' was also shown as lines added in 'Changes Drawer'.
The cause of this issue was it was comparing working copy with original unformatted content. This change fixes the issue by comparing current content with original formatted content
if the working copy is a formatted content. Since, formatting is expensive original formatted content is cached here.

This change addresses two cases of minified content changes.
1. Format minified content then, change.
   In this case, initial formatted value is saved in _formattingComplete called from _formatSourceInPlace click event handler.
2. Change minified content then, format.
   In this case, the first formatted content has a change. So, it computes the formatted value from original content then, caches it in _innerRequestDiff.

Bug: 977337
Change-Id: Ie48c6d990401819aaae831a51b6b7ac4107dae11