[devtools] Add Url and Path Column to Network Log
This patch allows removing the name column from the network
pane as long as one of (name,path,url) columns is still
visible.

This patch also adds a columnIndex parameter to DataGrid Cells
to make them aware of their column index. This is needed for
the group by frame toggle in the network pane to be able to add
the groups header to the first visible column.

Bug: chromium:993366
Change-Id: I864bdae899bba14514105095ebccbe732d1f9427