DevTools: Enable default keyboarding for TreeOutlines
Currently, if a TreeOutline does not have a selected item, it cannot be
navigated with a keyboard. This change fixes this by adding a tab stop
to the tree outline if there is no selected item and the tree outline
is focusable (this does not change the keyboard order in the console).
Once focus is placed on the tree outline, [ArrowUp] will select the
last item, and [ArrowDown] will select the first item.
GIF: https://imgur.com/a/yqq33FL

This PR previously covered maintaining selection when filtering tree
items, but this has been pulled out into a separate CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1792785

Testing: Use the Layers sidebar.
Click in empty space in the layers canvas to deselect the tree item,
then use [Shift]+[Tab] to navigate back to the tree outline. Then use
the arrow keys to select as mentioned above.

Original CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1696383

Bug: 963183, 964075
Change-Id: Ibabd6fe3e1faea4093f42b4628e7e7833c7b4b00