DevTools: Maintain selection when tree items are hidden
Previously when filtering a tree, if the selected element was a child,
selection could be lost, preventing a keyboard user from being able to
navigate through the filtered list. This change fixes this problem in
the following ways:

1. If the selected tree item is hidden, the first item in the tree is
selected. This can be seen in the Computed Styles pane where after
filtering results, a user could no longer reach any items in the tree.

2. If focus is on a child element and the parent is collapsed, then the
parent is selected. This is only a problem if using BOTH keyboard and
mouse since focus is already moved to parent when using arrow keys.

Before and after: https://imgur.com/a/2pbqfrm

Original CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1792785

Bug: 964075, 963183
Change-Id: Ia8e44f5f13fdafa6819c1771fa25712402354fab