[bindings] Fix location translation for inline script with #sourceURL=
Translation between source locations of inline scripts with #sourceURL=
did not work correctly, because the SourceFormatter handled them like
inline scripts. DevTools opens them in its own window, and then
continues to refer to locations in the file by the original positions
in the containing document; however, the line endings used for
formatting are line endings for the script only (in contrast to line
endings for other inline scripts, which just reuse the line endings
of the containing document).

This CL adds a hack to handle inline scripts with #sourceURL= separately
which entails that the SourceFormatter now also needs to know that
such scripts are displayed in their own frame (as it is supposed to
produce valid uiLocations). A redesign at some point in the future must
clean this up; at the moment only fragmental knowledge about how this is
supposed to work exists.

Bug: chromium:1003497, chromium:1005789
Change-Id: I3e13c6bc76b8692ce6badc09880c41d1a7e444c4