[DevTools] Update SecurityPanel for grey triangle experiment
Security UX is experimenting with using a grey triangle warning for
non-secure site connections (crbug.com/997972). When this icon is
shown, DevTools Security panel's security overview should match the
state shown in the omnibox.

Previously, crrev.com/c/1874787 added the insecure-broken state to
the DevTools protocol, so this cl does the setup so that once the
insecure-broken state gets sent for the DANGEROUS security state
(crrev.com/c/1879498, which needs to be submitted after the frontend
changes) and non-secure connections get styled as kInsecure
(crrev.com/c/1877653), the triangle pointer will point to the correct
icon.

Bug: 1008218
Change-Id: Ib893489a9d7c450892eff5e33bbbf3f2aef2e283