ListControl ARIA improvements and opt-out
This CL marks UI.ListControls as listboxes and their list items as
options, in accordance with the WAI-ARIA listbox pattern at [1]. If the
element returned by createElementForItem already has a role,
UI.ListControl will not change it. This also adds a new
updateSelectedItemARIA() method to UI.ListDelegate in order to let
ListControl users opt out of the default management of aria-selected.

Other changes:
- calls preventDefault in the keydown handler to prevent arrowing
through the list from unnecessarily scrolling
- throws an error instead of a string

[1] https://www.w3.org/TR/wai-aria-practices/#Listbox

Bug: 963183
Change-Id: I2a81744371497d0e592556f65f0c8177826d7c19