DevTools: Make request headers keyboard navigable
Currently, the request headers tab cannot be manipulated with a
keyboard whatsoever. This means that a keyboard user cannot expand or
collapse the groups in the tree to focus on the content that they want
to see.

This change enables the keyboarding by allowing selection of the tree
and its nodes, by fixing the focus styling (so focus is always visible),
and by selecting the first node as an entry point for tabbing into the
content.

Bug: 963183
Change-Id: Ie48032436ef91b5d7de974a728f184021e20de1e