DevTools: Add keyboard navigation in Request Payload object
Previously, in the Request Payload section of the Headers tab in the
Network tool, an ObjectPropertiesSection was inserted as a tree within a
tree element, which prevented keyboaring in the object itself. This
means that keyboard users were blocked from being able to access the
content of the object since only the root node is expanded by default.

This change appends the root node of the object into the existing tree,
so users can navigate through the elements of the object just the same
as the rest of the items in the tree of the Headers tab.

Bug: 963183
Change-Id: I7df639f97d4f807e25af854fc8a16f097837a1e0