DevTools: Add context menu entries for Request Payload toggles
In the the Network Headers pane, the buttons for "View source/parsed"
were not accessible via keyboard. This change adds context menu entries
to the Request Payload list item, so a keyboard user can access these
toggles using a context menu key or shortcut.

Note: this does not fix the toggles for other sections in the headers
pane

Testing: Go to youtube, open Network tool and refresh the page. There
should be an entry with the name "log_event?alt=json&key=..." which has
a request payload for verifying this change.

Bug: 963183
Change-Id: I3c5807b66d65da2b1f0f1dbfa84e761036104851