DevTools: Add context menu entries for toggles in Query String Parameters
In the the Network Headers pane, the buttons for "View source/parsed"
and "View encoded/decoded" were not accessible via keyboard. This change
adds context menu entries to the Query String Paramers list item, so a
keyboard user can access these toggles using a context menu key or
shortcut.

Note: this does not fix the toggles for other sections in the headers
pane

Screenshot: https://imgur.com/a/N1DjlhQ

Bug: 963183
Change-Id: Ia04f97ecc0cf1d6a3bae74198a5d4a8d6f2cae0a