DevTools: Add context menu entry for "Show more" button
In the Network Headers pane long Request Payloads are truncated and have
a "Show more" button that allows a user to show the entire payload. This
button is not keyboard accessible.

This change adds a context menu entry to the tree item containing the
request payload, so that a keyboard user can access this action.

Bug: 963183
Change-Id: Id7e9fec6861be5627eff1007cd27643cf9c426f2