DevTools: Introducing devtools_disable_remote_modules flag (DT Repo)
Use the recently added devtools_disable_remote_modules flag
to include remote modules when the in the resources.pak file. By doing
this remote modules will be forced to be consumed from the pak file
which allows DevTools developers to test their local changes.

This CL requires:
https://chromium-review.googlesource.com/c/chromium/src/+/1880957

which lives in the chromium repository.

Change-Id: Ib3716018720f90a30ae6ee24b314171fa7e8543e