Network A11y: Improve keyboarding experience of Network tool Datagrid
This changes builds off of <INSERT NAME HERE> in order to improve
the keyboard accessibility of the Network tool Datagrid.

Previously, the network tool ties selection of a node in the grid
directly to opening the sidepane. This makes it challenging for
keyboard only users to view and navigate the network tool datagrid
columns.

This change separates selection from opening the sidepane, and fixes
styling issues that affect the columns that were previously hidden
behind the sidepane when a node is selected. Selected nodes are
highlighted in the grid, and are "activated" using the enter key
or a click in the primary column to open the sidepane.

Before:

After:
Change-Id: I2c2629670624e807adb4cd9e9e9b75cdafbae7a5