[coverage] Only support per function coverage
Per block coverage is far more expensive to collect, so we
opt for per function coverage for now. We plan on bringing
back block coverage at some point as an additional choice.

The main draw-back of the more coarse-grained per function
coverage is that inline scripts are now reported as covered
as soon as their first line was executed (because inline
scripts are treated like a function on the V8 side).

Here is a comparison. Note how the inline scripts in the
index file now show higher coverage.

Before:
After: https://imgur.com/a/v23Uk73

Bug: chromium:1009396
Change-Id: I1c0da7b45c099206028f08300d7c3a92b10a4433