Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > IIUC this would be overwritten once you do a `gclient sync` in `chromium/src`, would it not? Is there any way of blocking a `gclient sync` of overwriting? Moreover, if it is overwritten, can `gclient sync` be instructed to ignore that folder so that other directory syncing still happens?
> 
> Yes. Once overwritten, you could simply rerun rsync. Alternatively, you can specify a custom_deps in the .gclient file to have third_party/devtools-frontend/src point to None.

We have hooks for our gclient, can we maybe take advantage of that? We could hardcode a folder location and say: if you put it in here, we will be smart. E.g. you have `folder/chromium/src` and if `folder/devtools-frontend` exists, we are going to symlink in. Not sure how that will work on a Windows machine, but maybe we can find something for that use case as well.

Patch-set: 2
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>