Update patch set 1
Patch Set 1:

Responding to Rob's feedback on the old PR:
third_party/blink/renderer/devtools/front_end/resources/DatabaseQueryView.js:

PS10, Line 72:
It looks to me as if `index` might equal -2 sometimes.

Reply: It shouldn't, it will get reset in the next line in that case to this._queryResultsWithDataGrids.length - 1 (i.e. -1). But, you have a point, making this more clear by using an else in both blocks.

PS10, Line 81:
I'm not sure I understand this condition. If no query is selected, and the user presses the down-arrow, it seems to me that the natural thing to do would be to select the first item.

Reply: So (this._selectedQueryIndex !== -1) would mean in case a query is selected, go to the next one. If not just stay. If I understand correctly, with this suggestion you'd go to the top query result - which could be very old and unrelated, so I thought it would be better to go to the prompt instead. Let me know what you think.

PTAL Rob, thanks!

Patch-set: 1