Add main-thread scrolling reasons to 3d layers view in the inspector
This is the devtools-frontend patch of a two-patch series to add main-
thread scrolling reasons to the 3d layers view in the inspector. The
inspector has a feature for showing the 3d layers of the page under
More tools > Layers. This 3d layer view includes an option for "Slow
scroll rects", and this visualization did not include main-thread
scrolling reasons such as background-attachment: fixed. This patch adds
plumbing to add main-thread scrolling reasons to the "Slow scroll rects"
visualization.

This patch also fixes a bug in the string format used for the slow
scroll rects. It was previously printed as:
  [x] x [y] (at [width], [height])
With this patch it is printed as:
  [width] x [height] (at [x], [y])

Bug: 1015918
Change-Id: I3cff6414c5f18bb9ed14a0705221d8012ccc18d1