[CSS Overview] Reduce the eagerness of reset
We currently observe the InspectedURLChanged event to decide if the
CSS Overview needs to be reset, but this can be fired, it seems, for
changes internal to the page (seems like iframes, perhaps). As such
with this CL we confirm that the inspected URL itself has changed
before triggering the reset.

Change-Id: I32b721b66bd75875ecd61426573131407c455a4f