Modify PRESUBMIT to run localization scripts on all files for commit.
After this change PRESUBMIT will have the following behavior:

CheckChangeOnUpload
✓ - modified file with (column) extension will trigger verification

+--------------------------------+-----+---------+-------+-------------------+
|          Script name           | .js | .grd(p) | .json | any other file(*) |
+--------------------------------+-----+---------+-------+-------------------+
| check_localizable_resources.js | ✓   | ✓      | ✓     | X                 |
| check_localizability.js        | ✓   | ✓      | ✓     | X                 |
+--------------------------------+-----+---------+-------+-------------------+

CheckChangeOnCommit
✓ - modified file with (column) extension will trigger verification

+--------------------------------+-----+---------+-------+-------------------+
|          Script name           | .js | .grd(p) | .json | any other file(*) |
+--------------------------------+-----+---------+-------+-------------------+
| check_localizable_resources.js | ✓   | ✓      | ✓     | ✓                 |
| check_localizability.js        | ✓   | ✓      | ✓     | ✓                 |
+--------------------------------+-----+---------+-------+-------------------+

There is a corner case that will be created with this change (by design)
If a user introduces a localization error(s) by using bypass-hooks the next user
in the queue will hit this issue and will be blocked for check in until the
issue gets fixed(or by using bypass-hooks)

e.g
Bob bypasses a change with localization errors [missing description in grd(p)],
later that day Alice adds a comment to a .css file. When Alice try to commit
Alice will hit Bob error even if Alice did not modified that file.

Change-Id: Iba3ffcc761e8151f3305d6167cde0485c40123d1