Allow ExpandableContainerWidget users to focus their title
This CL allows ExpandableContainerWidget children to focus their header
elements when they have no focusable content of their own. This is
necessary because trying to open any of the Sources panel sidebar tools
(for example, global listeners) via the command menu will focus their
content if it exists, but otherwise just show the panel without moving
focus or providing any indication of what is happening to screen reader
users. This was originally going to be handled by focusing their empty
descriptions [1], but that received pushback.

Example of ExpandableContainerWidget children:
https://gyazo.com/f6f955d197ad05cf5321e8cb1669955f

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1717157/

Bug: 963183
Change-Id: I23c486d147f32661e3f6d6e23b6c01c07db2e1b3