DevTools: Make shades of a color palette keyboard accessible
Color swatches in Material palette category are special it has shades of color.
These shades of color appear when user clicks a color in that category for a longer period of time value used is 200ms.
That feature is made keyboard accessible in this change.

Keyboard accessiblity: In this change user can press ENTER for longer period around 200ms to get shades of color appears. Then, user can navigate through shades of a color using UP and DOWN arrow keys.
The value set for longer keypress is identical to the value set for longer mousedown which is 200ms.
Pressing ENTER or SPACEBAR in any of those focused shade of color will select that color.

Gif showing the changes https://imgur.com/nziRVJt

This change was previously reviewed here https://chromium-review.googlesource.com/c/chromium/src/+/1816873

Bug: 963183
Change-Id: I0356abb9e8b8cf5d1652d0e66861fae8d22a415b