Datagrid A11y: Select datagrid row on up/down arrow key press
This CL adds the ability to select the first or last entry in
a datagrid when it receives focus by pressing the up or down arrow
key. This change is to allow keyboard users to access the datagrid
without using a mouse.

Before: https://imgur.com/EkqXd95

After: https://imgur.com/fO3qp9W

Bug: 963183


Change-Id: I7b5ee2f2b5ff535bca9a2a10f5efe177c1322668