Implement lazy loading of modules
Any module that is not autostart and has `modules` specified in its `module.json`
will now have its entrypoint dynamically imported. Update the release build script
to output the `modules` array as well so that the Runtime can load it.

Add additional entrypoints to `platform` and `dom_extension` to be consistent
with the naming patterns of all other modules

Bug:1006759
Change-Id: If6d10a13a62354079e3f8ee49bee4ecdcffa6758