DevTools: Hide Network datagrid from SR when landing page is visible
Previously, when navigating the Network landing page with a screen
reader in browse/scan mode, the user would be able to reach the
DataGrid elements that are not visible. This content is reached before
reaching the landing page, so a user wouldn't know to keep navigating
past the grid to reach the shortcut suggestions.

This change hides the datagrid and toolbar from the screen reader using
the aria-hidden attribute, so focus moves straight to
the landing page content.

Original CL (reverted):
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1880306

Bug: 963183
Change-Id: I056e1a6aaafdcca86d22f358e99faa7e417d24bd