Scope Pane Keyboard Navigation
In order to enable keyboard navigation of the Scope pane, this CL moves
all of the Scope pane ObjectPropertiesSections into one
ObjectPropertiesSectionTreeOutline to allow users to navigate them with
the keyboard. Users can tab into the scope treeview and navigate through
it with the arrow keys. Editable values can be changed by pressing
enter, and other functionality can be accessed via the context menu.

These changes break two tests, so [1] must be merged first to disable
them. The tests will be fixed by [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1894014
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1811968

Bug: 963183
Change-Id: I93a5b8b7bccc152b33faa171540c7ed99af1e4a3