Update patch set 2
Patch Set 2:

We had a discussion about that and figured that in comparison to the amount of things like the RunTask or UpdateLayer events the usual amount of parseOnBackground* events is quite small.
We also tried an implementation that sent the intervals of parsing and waiting in a single event at the end but that had the disadvantage of not being able to show any information if the end event was missing.

Patch-set: 2